some refactor

+++ src/main/java/com/emptyqueue/backend/QueueController.java
@@ -29,25 +29,34 @@ public class QueueController {
             return ResponseEntity.badRequest().build();
         }
         return ResponseEntity.ok(queueRepository.findAllByOwner(principal.getName()));
-        // return queueRepository.findAllByOwner(principal.getName());
     }
-
+    
     @GetMapping("/{id}")
-    public Queue getQueue(@PathVariable Long id) {
-        return queueRepository.findById(id).orElseThrow(RuntimeException::new);
+    public ResponseEntity getQueue(@PathVariable Long id, Principal principal) {
+        var queue = queueRepository.findById(id).orElseThrow(RuntimeException::new);
+        if (principal == null || principal.getName() != queue.getOwner()) {
+            return ResponseEntity.badRequest().build();
+        }
+        return ResponseEntity.ok(queue);
     }
 
 
     @PostMapping
     public ResponseEntity createQueue(@RequestBody Queue queue, Principal principal) throws URISyntaxException {
+        if (principal == null) {
+            return ResponseEntity.badRequest().build();
+        }
         queue.setOwner(principal.getName());
         Queue savedQueue = queueRepository.save(queue);
         return ResponseEntity.created(new URI("/queue/" + savedQueue.getId())).body(savedQueue);
     }
 
     @PostMapping("/{id}")
-    public ResponseEntity editQueue(@PathVariable Long id, @RequestBody Queue queue) throws URISyntaxException {
-        Queue currentQueue = queueRepository.findById(id).orElseThrow(RuntimeException::new);
+    public ResponseEntity editQueue(@PathVariable Long id, @RequestBody Queue queue, Principal principal) throws URISyntaxException {
+        var currentQueue = queueRepository.findById(id).orElseThrow(RuntimeException::new);
+        if (principal == null || principal.getName() != currentQueue.getOwner()) {
+            return ResponseEntity.badRequest().build();
+        }
         currentQueue.setName(queue.getName());
         currentQueue.setParticipants(queue.getParticipants());
         currentQueue.setCurrent(0);
@@ -57,8 +66,11 @@ public class QueueController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity iterateQueue(@PathVariable Long id) {
+    public ResponseEntity iterateQueue(@PathVariable Long id, Principal principal) {
         Queue currentQueue = queueRepository.findById(id).orElseThrow(RuntimeException::new);
+        if (principal == null || principal.getName() != currentQueue.getOwner()) {
+            return ResponseEntity.badRequest().build();
+        }        
         currentQueue.iterate();
         currentQueue = queueRepository.save(currentQueue);
 
@@ -66,7 +78,11 @@ public class QueueController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity deleteQueue(@PathVariable Long id) {
+    public ResponseEntity deleteQueue(@PathVariable Long id, Principal principal) {
+        Queue currentQueue = queueRepository.findById(id).orElseThrow(RuntimeException::new);
+        if (principal == null || principal.getName() != currentQueue.getOwner()) {
+            return ResponseEntity.badRequest().build();
+        } 
         queueRepository.deleteById(id);
         return ResponseEntity.ok().build();
     }
